moved FrameInfo
