+++ project/source/game/stats/UnitData.h
@@ -11,6 +11,7 @@
 //
 #include "ItemScript.h"
 #include "UnitGroup.h"
+#include "FrameInfo.h"
 
 //-----------------------------------------------------------------------------
 struct Spell;
@@ -173,70 +174,6 @@ struct SoundPack
 };
 
 //-----------------------------------------------------------------------------
-// Typy ramek
-enum FRAME_INDEX
-{
-	F_CAST,
-	F_TAKE_WEAPON,
-	F_ATTACK1_START,
-	F_ATTACK1_END,
-	F_ATTACK2_START,
-	F_ATTACK2_END,
-	F_ATTACK3_START,
-	F_ATTACK3_END,
-	F_BASH,
-	F_MAX
-};
-
-//-----------------------------------------------------------------------------
-// Flags for which weapon use which attack animation
-#define A_SHORT_BLADE (1<<0)
-#define A_LONG_BLADE (1<<1)
-#define A_BLUNT (1<<2)
-#define A_AXE (1<<3)
-
-//-----------------------------------------------------------------------------
-// Informacje o animacjach ataku postaci
-struct AttackFrameInfo
-{
-	struct Entry
-	{
-		float start, end;
-		int flags;
-
-		float Lerp() const
-		{
-			return ::Lerp(start, end, 2.f / 3);
-		}
-	};
-	vector<Entry> e;
-};
-
-//-----------------------------------------------------------------------------
-// Informacje o ramce animacji
-struct FrameInfo
-{
-	string id;
-	AttackFrameInfo* extra;
-	float t[F_MAX];
-	int attacks;
-
-	FrameInfo() : extra(nullptr), attacks(0), t() {}
-	~FrameInfo()
-	{
-		delete extra;
-	}
-
-	float Lerp(int frame) const
-	{
-		return ::Lerp(t[frame], t[frame + 1], 2.f / 3);
-	}
-
-	static vector<FrameInfo*> frames;
-	static FrameInfo* TryGet(const AnyString& id);
-};
-
-//-----------------------------------------------------------------------------
 struct IdlePack
 {
 	string id;
@@ -311,19 +248,8 @@ struct UnitData
 
 	void CopyFrom(UnitData& ud);
 
+	static SetContainer<UnitData> units;
 	static std::map<string, UnitData*> aliases;
 	static UnitData* TryGet(const AnyString& id);
 	static UnitData* Get(const AnyString& id);
 };
-
-//-----------------------------------------------------------------------------
-struct UnitDataComparer
-{
-	bool operator() (const UnitData* ud1, const UnitData* ud2) const
-	{
-		return _stricmp(ud1->id.c_str(), ud2->id.c_str()) > 0;
-	}
-};
-typedef std::set<UnitData*, UnitDataComparer> UnitDataContainer;
-typedef UnitDataContainer::iterator UnitDataIterator;
-extern UnitDataContainer unit_datas;
